requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain-google-genai
langchain-community
PyPDF2
faiss-cpu
sentence-transformers
